requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
alabaster=0.7.12=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
babel=2.9.1=pyhd3eb1b0_0
blas=1.0=mkl
bottleneck=1.3.4=py39h67323c0_0
brotlipy=0.7.0=py39h9ed2024_1003
bzip2=1.0.8=h1de35cc_0
ca-certificates=2022.3.18=hecd8cb5_0
certifi=2021.10.8=py39hecd8cb5_2
cffi=1.15.0=py39hc55c11b_1
charset-normalizer=2.0.12=pypi_0
colorama=0.4.4=pyhd3eb1b0_0
cpuonly=2.0=0
cryptography=36.0.0=py39hf6deb26_0
dgl=0.8.0post1=py39_0
docutils=0.17.1=py39hecd8cb5_1
ffmpeg=4.3=h0a44026_0
freetype=2.11.0=hd8bbffd_0
gettext=0.21.0=h7535e17_0
giflib=5.2.1=haf1e3a3_0
gmp=6.2.1=h23ab428_2
gnutls=3.6.15=hed9c0bf_0
icu=58.2=h0a44026_3
idna=3.3=pyhd3eb1b0_0
imagesize=1.3.0=pyhd3eb1b0_0
intel-openmp=2021.4.0=hecd8cb5_3538
jinja2=3.1.1=pypi_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h9ed2024_0
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libidn2=2.3.2=h9ed2024_0
libpng=1.6.37=ha441bb4_0
libtasn1=4.16.0=h9ed2024_0
libtiff=4.2.0=h87d7836_0
libunistring=0.9.10=h9ed2024_0
libuv=1.40.0=haf1e3a3_0
libwebp=1.2.2=h56c3ce4_0
libwebp-base=1.2.2=hca72f7f_0
libxml2=2.9.12=hcdb78fc_0
livereload=2.6.3=pyh9f0ad1d_0
llvm-openmp=12.0.0=h0dcd299_1
llvmlite=0.38.0=pypi_0
lz4-c=1.9.3=h23ab428_1
markdown-it-py=1.1.0=pyhd8ed1ab_0
markupsafe=2.1.1=pypi_0
mdit-py-plugins=0.2.8=pyhd8ed1ab_0
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py39h9ed2024_0
mkl_fft=1.3.1=py39h4ab4a9b_0
mkl_random=1.2.2=py39hb2f4e1b_0
myst-parser=0.15.2=pyhd8ed1ab_0
ncurses=6.3=hca72f7f_2
nettle=3.7.3=h230ac6f_1
networkx=2.7.1=pyhd3eb1b0_0
numba=0.55.1=pypi_0
numexpr=2.8.1=py39h2e5f0a9_0
numpy=1.21.5=pypi_0
numpy-base=1.21.2=py39he0bd621_0
openh264=2.1.1=h8346a28_0
openssl=1.1.1n=hca72f7f_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.1=pypi_0
pillow=9.0.1=py39hde71d04_0
pip=21.2.4=py39hecd8cb5_0
pycparser=2.21=pyhd3eb1b0_0
pyg=2.0.4=py39_torch_1.11.0_cpu
pygcl=0.1.2=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.7=pypi_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.11=hdfd78df_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python-louvain=0.15=pyhd3eb1b0_0
pytorch=1.11.0=py3.9_0
pytorch-cluster=1.6.0=py39_torch_1.11.0_cpu
pytorch-mutex=1.0=cpu
pytorch-scatter=2.0.9=py39_torch_1.11.0_cpu
pytorch-sparse=0.6.13=py39_torch_1.11.0_cpu
pytorch-spline-conv=1.2.1=py39_torch_1.11.0_cpu
pytz=2022.1=pypi_0
pyyaml=6.0=py39hca72f7f_1
readline=8.1.2=hca72f7f_1
requests=2.27.1=pyhd3eb1b0_0
scikit-learn=1.0.2=py39hae1ba45_1
scipy=1.8.0=pypi_0
setuptools=58.0.4=py39hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
snowballstemmer=2.2.0=pyhd3eb1b0_0
sphinx=4.0.2=pyhd3eb1b0_0
sphinx-autodoc-typehints=1.12.0=pyhd8ed1ab_0
sphinx_rtd_theme=1.0.0=pyhd8ed1ab_0
sphinxcontrib-applehelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-devhelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-htmlhelp=2.0.0=pyhd3eb1b0_0
sphinxcontrib-jsmath=1.0.1=pyhd3eb1b0_0
sphinxcontrib-qthelp=1.0.3=pyhd3eb1b0_0
sphinxcontrib-serializinghtml=1.1.5=pyhd3eb1b0_0
sqlite=3.37.2=h707629a_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.11=h7bc2e8c_0
torch=1.11.0=pypi_0
torchvision=0.12.0=py39_cpu
tornado=6.1=py39h9ed2024_0
tqdm=4.63.1=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2021e=hda174b7_0
urllib3=1.26.9=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
yacs=0.1.6=pyhd3eb1b0_1
yaml=0.2.5=haf1e3a3_0
zlib=1.2.11=h4dc903c_4
zstd=1.4.9=h322a384_0
